Separate map from the rest of the code
This uses the macro that Franco wrote to read our custom map format. Somehow I was supposed to convert the array of strings into an enumeration that tracks the appropriate tile for each map cell, but I'm not quite sure how that works. This definitely isn't what Franco intended, but it gets the job done for now.