Add a beautiful parallax background (fixes #14)
OK, it could definitely still use some work. The white tiles against a mostly-white background is confusing. The background is oubviously not filling the full space when you zoom out. If you move too far to the right, you start to run out of background. Etc, etc. We'll open issues for these things. It's a start.